pyproject.toml
[tool.poetry]
name = "ai-solutoin-architect"
version = "0.1.0"
description = "Crew that finds solution to a given architecture problem"
authors = ["Rakeh Gohel"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = "^0.22.5"
unstructured = '==0.10.25'
python-dotenv = "1.0.0"
langchain_openai = "0.0.5"
langchain-anthropic = "^0.1.6"

[tool.poetry.group.dev.dependencies]
deptry = "^0.16.1"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"